hf-space/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.9.2
pydantic-settings==2.3.4
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
psycopg2-binary==2.9.9
better-exceptions==0.3.3
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
slowapi==0.1.9
asyncpg==0.30.0
python-socketio==5.11.4
nest-asyncio==1.6.0
openai==1.52.2
mcp==1.1.2
PyJWT==2.8.0
email-validator==2.1.0